requirements.txt
pyyaml
jinja2
feedparser